# Adds flags for compiler optimizations