--- a/make/compiler_flags
+++ b/make/compiler_flags
@@ -74,7 +74,7 @@ ifneq (,$(findstring mingw32-g,$(CXX_TYPE)))
 else ifneq (,$(findstring gcc,$(CXX_TYPE)))
 	#CXXFLAGS_OPTIM ?= -fweb -fivopts -ftree-loop-linear -fstrict-aliasing -funroll-loops -floop-block -floop-nest-optimize
 	ifeq ($(shell expr $(CXX_MAJOR) \>= 5), 1)
-		CXXFLAGS_FLTO ?= -flto=jobserver
+		CXXFLAGS_FLTO ?= -flto=jobserver -fdevirtualize-speculatively -fdevirtualize-at-ltrans -fweb -fivopts -ftree-loop-linear -fstrict-aliasing -funroll-loops -floop-block -floop-nest-optimize
 		#CXXFLAGS_OPTIM += -fdevirtualize-speculatively -fdevirtualize-at-ltrans
 	endif
 endif