--- a/make/compiler_flags
+++ b/make/compiler_flags
@@ -65,16 +65,15 @@ CXX_MINOR := $(shell $(CXX) -dumpversion 2>&1 | cut -d'.' -f2)
 ifneq (,$(findstring clang,$(CXX_TYPE)))
 	CXXFLAGS_OPTIM ?= -fvectorize -ftree-vectorize -fslp-vectorize -ftree-slp-vectorize
 	ifeq ($(shell expr $(CXX_MAJOR) \>= 5), 1)
-		CXXFLAGS_FLTO ?= -flto=full -fwhole-program-vtables -fstrict-vtable-pointers
+		CXXFLAGS_FLTO ?= -flto=full -fwhole-program-vtables -fstrict-vtable-pointers -fforce-emit-vtables
 		AR := llvm-ar
-		CXXFLAGS_OPTIOM += -fforce-emit-vtables
 	endif
 endif
 ifneq (,$(findstring mingw32-g,$(CXX_TYPE)))
 else ifneq (,$(findstring gcc,$(CXX_TYPE)))
 	#CXXFLAGS_OPTIM ?= -fweb -fivopts -ftree-loop-linear -fstrict-aliasing -funroll-loops -floop-block -floop-nest-optimize
 	ifeq ($(shell expr $(CXX_MAJOR) \>= 5), 1)
-		CXXFLAGS_FLTO ?= -fdevirtualize-speculatively -fdevirtualize-at-ltrans -fweb -fivopts -ftree-loop-linear -fstrict-aliasing -funroll-loops -floop-block -floop-nest-optimize
+		CXXFLAGS_FLTO ?= -flto -fuse-linker-plugin -flto-partition=1to1 -fdevirtualize-speculatively -fdevirtualize-at-ltrans -fweb -fivopts -ftree-loop-linear -fstrict-aliasing -funroll-loops -floop-block -floop-nest-optimize
 		#CXXFLAGS_OPTIM += -fdevirtualize-speculatively -fdevirtualize-at-ltrans
 	endif
 endif