--- a/make/compiler_flags
+++ b/make/compiler_flags
@@ -63,7 +63,7 @@ CXX_MINOR := $(shell $(CXX) -dumpversion 2>&1 | cut -d'.' -f2)
 
 
 ifneq (,$(findstring clang,$(CXX_TYPE)))
-	CXXFLAGS_OPTIM ?= -fvectorize -ftree-vectorize -fslp-vectorize -ftree-slp-vectorize -fno-standalone-debug -fstrict-return -ftrigraphs -funroll-loops
+	CXXFLAGS_OPTIM ?= -fvectorize -ftree-vectorize -fslp-vectorize -ftree-slp-vectorize -fno-standalone-debug -fstrict-return -ftrigraphs -funroll-loops -fvariable-expansion-in-unroller
 	ifeq ($(shell expr $(CXX_MAJOR) \>= 5), 1)
 		CXXFLAGS_FLTO ?= -flto=full -fwhole-program-vtables -fstrict-vtable-pointers -fforce-emit-vtables
 		AR := llvm-ar-$(CXX_MAJOR)
@@ -71,13 +71,13 @@ ifneq (,$(findstring clang,$(CXX_TYPE)))
 endif
 ifneq (,$(findstring mingw32-g,$(CXX_TYPE)))
 else ifneq (,$(findstring gcc,$(CXX_TYPE)))
-	CXXFLAGS_OPTIM ?= -fweb -fivopts -ftree-loop-linear -fstrict-aliasing -funroll-loops -floop-block -floop-nest-optimize -ftree-loop-distribution -ftree-loop-if-convert-stores -ftree-vectorize
+	CXXFLAGS_OPTIM ?= -fweb -fivopts -ftree-loop-linear -funroll-loops -floop-block -floop-nest-optimize -ftracer
+	ifeq ($(shell expr $(CXX_MAJOR) \>= 5), 1)
+	  CXXFLAGS_OPTIM += -fdevirtualize-speculatively -fdevirtualize-at-ltrans
+  endif
 	ifeq ($(shell expr $(CXX_MAJOR) \> 9), 1)
 		CXXFLAGS_FLTO ?= -flto=jobserver -fuse-linker-plugin
 	endif
-	ifeq ($(shell expr $(CXX_MAJOR) \>= 5), 1)
-	  CXXFLAGS_OPTIM += -fdevirtualize-speculatively -fdevirtualize-at-ltrans -fgcse-sm -fgcse-las -fgcse-after-reload 
-  endif
 endif
 
 ################################################################################