--- a/make/compiler_flags
+++ b/make/compiler_flags
@@ -63,7 +63,7 @@ CXX_MINOR := $(shell $(CXX) -dumpversion 2>&1 | cut -d'.' -f2)
 
 
 ifneq (,$(findstring clang,$(CXX_TYPE)))
-	CXXFLAGS_OPTIM ?= -fvectorize -ftree-vectorize -fslp-vectorize -ftree-slp-vectorize -fno-standalone-debug -fstrict-return -ftrigraphs -funroll-loops -fvariable-expansion-in-unroller
+	CXXFLAGS_OPTIM ?= -fvectorize -ftree-vectorize -fslp-vectorize -ftree-slp-vectorize -fno-standalone-debug -fstrict-return -ftrigraphs
 	ifeq ($(shell expr $(CXX_MAJOR) \>= 5), 1)
 		CXXFLAGS_FLTO ?= -flto=full -fwhole-program-vtables -fstrict-vtable-pointers -fforce-emit-vtables
 		AR := llvm-ar-$(CXX_MAJOR)