## Summary

After adding a template parameter to `vari` we saw a 20% regression in the SIR model. This PR includes some extra compiler flag options so if a user has either `clang >= 8` or `gcc >= 5`. flto and other compiler optimization flags will turn on. flto will only turn on for `clang >= 8` because clang-6 and clang-7's `-dumpversion` give `4.2.1` so we cannot tell them apart. `clang-6` requires the use of either the `ar` gold linker plugin or `llvm-ar`. Since the gold plugin is a non-standard extension for some versions of `ar` to have installed we turn off flto by default for both clang 6 and 7

## Tests

make feature so no new tests, though because our benchmarks use clang-6 they won't give the new benchmark values so I will paste them here. These tests were run on Ubuntu 20.04 with AMD Threadripper 1950x along with DDR4 ram. Tests for each compiler were run 10 times then averaged together

#### Clang-10 

The clang-10 results were run with 5000 warmup and 5000 samples, though overall I didn't see the increase in iteration numbers help produce better reliability in the test numbers.

| Name | Old Result | New Result | Ratio | Performance change( 1 - new / old ) |
| ------------- |------------- | ------------- | ------------- | ------------- |
| gp_pois_regr/gp_pois_regr.stan | 7.1 | 6.71 | 1.06 | 5.49% faster |
| low_dim_corr_gauss/low_dim_corr_gauss.stan | 0.01 | 0.02 | 0.88 | -13.0% slower |
| eight_schools/eight_schools.stan | 0.18 | 0.17 | 1.03 | 3.28% faster |
| arK/arK.stan | 5.27 | 5.28 | 1.0 | -0.16% slower |
| irt_2pl/irt_2pl.stan | 15.57 | 15.87 | 0.98 | -1.88% slower |
| performance.compilation | 48.23 | 58.49 | 0.82 | -21.28% slower |
| low_dim_gauss_mix_collapse/low_dim_gauss_mix_collapse.stan | 23.93 | 22.65 | 1.06 | 5.38% faster |
| pkpd/one_comp_mm_elim_abs.stan | 59.22 | 56.66 | 1.05 | 4.31% faster |
| sir/sir.stan | 224.85 | 190.03 | 1.18 | 15.48% faster |
| pkpd/sim_one_comp_mm_elim_abs.stan | 0.25 | 0.25 | 1.01 | 0.75% faster |
| low_dim_gauss_mix/low_dim_gauss_mix.stan | 6.93 | 6.61 | 1.05 | 4.66% faster |
| gp_regr/gp_regr.stan | 0.33 | 0.32 | 1.03 | 2.54% faster |
| arma/arma.stan | 0.76 | 0.75 | 1.02 | 1.96% faster |
| gp_regr/gen_gp_data.stan | 0.03 | 0.03 | 1.0 | 0.24% faster |
| garch/garch.stan | 1.97 | 1.75 | 1.13 | 11.35% faster |
Mean result: 1.0199644747


#### GCC-10

Because I didn't see any increased reliability from running more than the standard number of iterations the gcc results were run with 1000 warmup and 1000 samples. 

| Name | Old Result | New Result | Ratio | Performance change( 1 - new / old ) |
| ------------- |------------- | ------------- | ------------- | ------------- |
| gp_pois_regr/gp_pois_regr.stan | 2.34 | 1.98 | 1.18 | 15.29% faster |
| low_dim_corr_gauss/low_dim_corr_gauss.stan | 0.01 | 0.01 | 1.05 | 4.75% faster |
| eight_schools/eight_schools.stan | 0.05 | 0.05 | 0.99 | -0.91% slower |
| arK/arK.stan | 1.24 | 1.25 | 0.99 | -0.95% slower |
| irt_2pl/irt_2pl.stan | 4.22 | 4.03 | 1.05 | 4.43% faster |
| performance.compilation | 86.27 | 96.83 | 0.89 | -12.23% slower |
| low_dim_gauss_mix_collapse/low_dim_gauss_mix_collapse.stan | 4.59 | 4.7 | 0.98 | -2.53% slower |
| pkpd/one_comp_mm_elim_abs.stan | 14.17 | 14.03 | 1.01 | 0.95% faster |
| sir/sir.stan | 87.9 | 70.69 | 1.24 | 19.57% faster |
| pkpd/sim_one_comp_mm_elim_abs.stan | 0.23 | 0.23 | 1.0 | 0.11% faster |
| low_dim_gauss_mix/low_dim_gauss_mix.stan | 1.68 | 1.7 | 0.99 | -1.08% slower |
| gp_regr/gp_regr.stan | 0.09 | 0.1 | 0.93 | -8.02% slower |
| arma/arma.stan | 0.33 | 0.34 | 0.98 | -1.56% slower |
| gp_regr/gen_gp_data.stan | 0.02 | 0.02 | 0.99 | -1.37% slower |
| garch/garch.stan | 0.34 | 0.37 | 0.91 | -9.7% slower |
Mean result: 1.01176292109

## Side Effects

There's are some compiler flags which require the gcc graphite compiler extension, though tmk this comes with most standard builds of gcc unless you build gcc from source in which case I believe you need to specify whether the graphite extension should be compiled or not. If a user does not have graphite setup a warning is given and the program continues to run.

One of the interesting thing I found out is that several of these tests have a pretty wide variance. For instance here is the results of running develop against develop 10 times

---RESULTS---
| Name | Old Result | New Result | Ratio | Performance change( 1 - new / old ) |
| ------------- |------------- | ------------- | ------------- | ------------- |
| gp_pois_regr/gp_pois_regr.stan | 2.34 | 2.29 | 1.02 | 2.37% faster |
| low_dim_corr_gauss/low_dim_corr_gauss.stan | 0.01 | 0.01 | 0.92 | -8.15% slower |
| eight_schools/eight_schools.stan | 0.05 | 0.05 | 1.01 | 0.92% faster |
| arK/arK.stan | 1.27 | 1.3 | 0.98 | -1.86% slower |
| irt_2pl/irt_2pl.stan | 4.23 | 4.18 | 1.01 | 1.24% faster |
| performance.compilation | 60.81 | 60.2 | 1.01 | 1.0% faster |
| low_dim_gauss_mix_collapse/low_dim_gauss_mix_collapse.stan | 4.63 | 4.79 | 0.97 | -3.4% slower |
| pkpd/one_comp_mm_elim_abs.stan | 14.07 | 14.15 | 0.99 | -0.57% slower |
| sir/sir.stan | 82.19 | 81.9 | 1.0 | 0.36% faster |
| pkpd/sim_one_comp_mm_elim_abs.stan | 0.25 | 0.25 | 1.03 | 2.95% faster |
| low_dim_gauss_mix/low_dim_gauss_mix.stan | 1.71 | 1.73 | 0.99 | -1.28% slower |
| gp_regr/gp_regr.stan | 0.1 | 0.1 | 1.01 | 1.36% faster |
| arma/arma.stan | 0.35 | 0.34 | 1.02 | 2.22% faster |
| gp_regr/gen_gp_data.stan | 0.03 | 0.03 | 1.05 | 5.15% faster |
| garch/garch.stan | 0.35 | 0.34 | 1.01 | 1.03% faster |


So `low_dim_corr_gauss` and `gen_gp_data` can have pretty large variance with even a decent number of runs


## Release notes

Add compiler flags to give better optimizations

## Checklist

- [x] Closes #2008 

- [x] Copyright holder: Steve Bronder

    The copyright holder is typically you or your assignee, such as a university or company. By submitting this pull request, the copyright holder is agreeing to the license the submitted work under the following licenses:
      - Code: BSD 3-clause (https://opensource.org/licenses/BSD-3-Clause)
      - Documentation: CC-BY 4.0 (https://creativecommons.org/licenses/by/4.0/)

- [x] the basic tests are passing

    - unit tests pass (to run, use: `./runTests.py test/unit`)
    - header checks pass, (`make test-headers`)
    - dependencies checks pass, (`make test-math-dependencies`)
    - docs build, (`make doxygen`)
    - code passes the built in [C++ standards](https://github.com/stan-dev/stan/wiki/Code-Quality) checks (`make cpplint`)

- [x] the code is written in idiomatic C++ and changes are documented in the doxygen

- [x] the new changes are tested