@rok-cesnovar 2020-08-18
We also need a quick summary to understand in which cases there is a regression. If I understand correctly then the logic is:
```
if g++ version < 5 or clang version < 8
  - some regression is expected for models where memory allocation is a big part of the model. Which I guess is ODE models? Any others?
else
  - no changes expected?
```