@wds15 2020-08-18
Just as FYI: I did find performance regression for SIR, yes, but also for a mixture logistic regression model!

Thanks for putting this together, it looks like a lot of work! I expected that this would be quick and easy, but it‘s a lot of work as it looks.