@rok-cesnovar 2020-08-18
> Just as FYI: I did find performance regression for SIR, yes, but also for a mixture logistic regression model!

Oh, thanks for the info. Then it will be a bit more challenging to get an intuition on.

IMHO g++ < 5 and clang < 8 is not going to be a lot our user base. Its mostly going to be RTools 3.5 Windows users, which hopefully will upgrade sooner rather than later. Anyone that cares much about performance and execution times has probably already upgraded to RTools 4.0. Last 3 Ubuntu LTS come with a compiler version >= 5.

I am not so sure on Macs, but if I am reading [this](https://trac.macports.org/wiki/XcodeVersionInfo) right, the supported clang versions in XCode for anything above and include Sierra is fine for us. I dont think there are many users still left on El Capitan and older.