@SteveBronder 2020-08-18
> Thanks! The performance tests look good.

np!

> I am a bit confused as to which clang versions allow these flags. In the Summary you mention clang 8 but in the side effects you write: "the most important one being that if a user uses a clang-[7-10] they must also have llvm-ar-[7-10] hooked into their path as well for flto.". I think you mention that there is some confusion with clang and versions anyways?

Yes sorry I submitted this before bed. This was a previous goof. If the user is using clang > 7 then we can use the system `ar`. I'll copy/paste what I wrote above

So this is a clang oddity. clang-6 and clang-7 both give 4.2.1 for -dumpversion. using -v instead of -dumpversion can give some weird results across operating systems and requires some extra trickery and gotchas to figure out. So while we could support flto by default with clang-7, clang-6 needs the llvm-ar. And since we can't tell them apart I think it's just easier to only support flto by default for systems with clang > 7.

> Another question is regarding "they must also have llvm-ar-[7-10] hooked into their path". What does this mean for a typical Stan user that is not an expert on C++ toolchains. Is this installed for Mac users with the Xcode toolchain and works out of the box? Does one have to do anything special on Ubuntu apart from "sudo apt install clang".

My b you can ignore that

> I am not as worried for anyone that compiles gcc on their own. Those that do, do not require much assistance wrt to setting up a working toolchain anyways.


That's what I figure as well

>> One of the interesting thing I found out is that several of these tests have a pretty wide variance.

> Yep, when doing some performance benchmarks in the past, I had to run like 40000 iterations for garch, gp_regr, eight_schools, and other quick models to get consistent timings.

Ugh yeah it makes me question whether we should really use those for benchmarks. Maybe we need to change up how we do the performance tests to something like how google perf does them where they run N times till they get a stable estimate.