@SteveBronder 2020-08-18
> Just as FYI: I did find performance regression for SIR, yes, but also for a mixture logistic regression model!

> Thanks for putting this together, it looks like a lot of work! I expected that this would be quick and easy, but it‘s a lot of work as it looks.

Yes these flags should fix up both of those cases. I can look at that logistic regression you posted as well. May be a good idea to add it to the perf tests.

> IMHO g++ < 5 and clang < 8 is not going to be a lot our user base. Its mostly going to be RTools 3.5 Windows users, which hopefully will upgrade sooner rather than later. Anyone that cares much about performance and execution times has probably already upgraded to RTools 4.0. Last 3 Ubuntu LTS come with a compiler version >= 5.

Yep!

> I am not so sure on Macs, but if I am reading this right, the supported clang versions in XCode for anything above and include Sierra is fine for us. I dont think there are many users still left on El Capitan and older.

Is our testing mac on El Capitan or Sierra? I know it has clang-7 as the default compiler so it's either an updated El Capitan or an outdated Sierra