@rok-cesnovar 2020-08-19
> Oh I just realized this is PR #2020
   Let's quadruple check this one so it doesn't blow up and ruin everything

Or let's just merge and get over it ASAP :)

Joking aside, this looks good now, there just seems to be a Windows related issue at the [Cmdstan level](https://jenkins.mc-stan.org/blue/organizations/jenkins/CmdStan/detail/downstream_tests/1925/pipeline). Will see if I can debug it a bit more.