@SteveBronder 2020-08-20
Ah, mingw's gcc ran into [this](https://jenkins.mc-stan.org/blue/organizations/jenkins/CmdStan/detail/downstream_tests/1925/pipeline) compiler bug. This was fixed in gcc-8, just to be safe I'm not turning on flto for anything less than gcc-8 because I couldn't tell from the discussion whether this patch was backported to earlier gcc versions