@rok-cesnovar 2020-08-20
Not sure I understand. This Windows machine uses g++ 8.3.0. How does setting the limit to 7 help here?