@SteveBronder 2020-08-20
Oh shoot I thought that part of the testing used Rtools 3.5. I'll have to dive into that more to figure out what's going on then