@SteveBronder 2020-08-20
Hmm okay [this](https://github.com/RIOT-OS/RIOT/issues/5775) hints that the issue is when a linked library is not compiled with flto but other parts are. I'll have to play around with this