@rok-cesnovar 2020-08-20
All Jenkins agents are now running RTools 4.0. The RTools 3.5 tests were moved to Github Actions. There will be an overview doc of all of this once the CI refactor is done.