@rok-cesnovar 2020-08-20
Hm, ok, then it might be the boost program options we build in cmdstan and statically link. Not sure where this pops up. Will take a look on my Windows machine.