@rok-cesnovar 2020-08-20
Can confirm the suspicion. You can build the model, the problem is `make build` that also builds bin/stansummary, which requires Boost program options. Stansummary is statically linked in. I presume that lib needs `-flto` and the rest of flags?

I am guessing you need to add the flags here: https://github.com/stan-dev/cmdstan/blob/develop/make/command
Let me know if I can help/test anything.