@SteveBronder 2020-08-20
@rok-cesnovar what do you think about just adding the makevar variables in this PR then in cmdstan adding these optimization flags? I think doing these flags upstream gives us a larger view on what we want to use flto on. I also don't totally trust windows and would bet a dollar we'll have to do something particular for rtools4.0 and pystan