@rok-cesnovar 2020-08-20
Yes, that seems like a good plan. Have them defined in math and use thevin cmdstan seems fine to me.