@rok-cesnovar 2020-08-23
Cool. The only thing I am a bit worried about is math performance tests won't have these flags now that they will live in Cmdstan. But I guess that can be handled separately?