@SteveBronder 2020-08-23
I thought the performance tests ran using the develop version of cmdstan, is that false?