@rok-cesnovar 2020-08-23
I mean if someone will do performance tests with Math-only (with Google benchmark for example), that won't be representative of how it will actually run in the context of Cmdstan. 

rstan/pystan do not use math makefiles anyways so its a different story there.