@SteveBronder 2020-08-23
Yeah thats true tho imo I think thats fine. Still technically a regression in the math lib but upstream performance improvement